Require auth when entering background

+++ Oak/Views/Settings/EditAccounts/EditAccountsViewModel.swift
@@ -7,40 +7,54 @@
 
 import SwiftUI
 import RealmSwift
+import Resolver
 
 class EditAccountsViewModel: ObservableObject {
-    private let accountService: AccountService
+    @Injected private var accountService: AccountService
+    
     private var accountsForDeletion: IndexSet?
-    private var accounts: Results<Account>
+    private var accounts: Results<Account>?
     private var realmToken: NotificationToken?
     
     @Published var accountRowModels: [AccountRowViewModel] = []
     @Published var showingConfirmDeletionAlert = false
     
-    init(accountService: AccountService) {
-        self.accountService = accountService
-        
+    deinit {
+        realmToken?.invalidate()
+    }
+    
+    func fetchAccounts() {
         // When our view model is initalised we'll setup our realm observer
         accounts = accountService.fetch()
         initializeRealmObserver()
     }
     
-    deinit {
-        realmToken?.invalidate()
-    }
-    
     /// Watch for changes to our accounts
     func initializeRealmObserver() {
-        realmToken = accounts.observe { [weak self] (changes: RealmCollectionChange<Results<Account>>) in
+        if let token = realmToken {
+            token.invalidate()
+        }
+        
+        realmToken = accounts?.observe { [weak self] (changes: RealmCollectionChange<Results<Account>>) in
             guard let self = self else {
                 return
             }
-            self.accountRowModels = self.accounts.map { AccountRowViewModel(account: $0) }
+            if let accounts = self.accounts {
+                self.accountRowModels = accounts.map { AccountRowViewModel(account: $0) }
+            }
         }
     }
     
     func deleteAccount(index: IndexSet) {
-        let items = index.map({ accounts[$0] })
-        try? accountService.delete(accounts: items)
+        if let accounts = accounts {
+            let items = index.map({ accounts[$0] })
+            try? accountService.delete(accounts: items)
+        }
+    }
+}
+
+extension Resolver {
+    static func RegisterEditAccountsViewModel() {
+        register { EditAccountsViewModel() }
     }
 }
